Correctly parse Dominion CVRs with multiple "Cards" in one "Session"
The Dominion JSON format has a Sessions key at the top level that
contains an array of objects. Each of these objects has a .Cards
property that is typically an array of one object that corresponds to a
CVR. But sometimes (~1-2% of the time) it's an array of multiple
objects, each of which is a CVR. Our Dominion CVR parser was ignoring
these, because it only ever looked at the first element of the array.

Instead, iterate over this array, so that we also capture the CVRs
stored in the second or third elements of these Cards arrays. Without
this change, some votes are missing when parsing San Francisco's
November 2020 data.